refactor(iota-types): move the consensus committee conversion to iota-core
`iota-types` depended on `starfish-config` for a single trait method,
`EpochStartSystemStateTrait::get_consensus_committee`, which converts an
epoch start state into a consensus `Committee`. That edge pulled
`starfish-config -> iota-network-stack -> iota-http -> iota-tls` under
`iota-types`, which nearly every crate in the workspace waits for.

- replace the trait method with a `committee_validators` accessor, since
  the validator list was private and the conversion needs it
- add an `EpochStartConsensusCommittee` extension trait in `iota-core`,
  which already depends on `starfish-config`, and move the conversion
  there unchanged
- move the two tests covering the IOTA/consensus committee ordering
  agreement along with it, constructing the state through the public
  `new_v1`/`new_v2` instead of private struct literals
- drop the `starfish-config` dependency from `iota-types`

All five call sites were already in `iota-core`, so no crate gains a new
dependency. `iota-types` drops from build level 4 to 3 and its 67
transitive dependents each shift down with it, shortening the longest
compilation chain in the workspace from 17 levels to 16.
