+++ crates/iota-core/src/consensus_manager/starfish_manager.rs
@@ -9,10 +9,7 @@ use async_trait::async_trait;
 use fastcrypto::ed25519;
 use iota_config::NodeConfig;
 use iota_metrics::{RegistryID, RegistryService, monitored_mpsc::unbounded_channel};
-use iota_types::{
-    committee::EpochId,
-    iota_system_state::epoch_start_iota_system_state::EpochStartSystemStateTrait,
-};
+use iota_types::committee::EpochId;
 use starfish_config::{Committee, NetworkKeyPair, Parameters, ProtocolKeyPair};
 use starfish_core::{
     Clock, CommitConsumer, CommitConsumerMonitor, CommitIndex, ConsensusAuthority,
@@ -27,6 +24,7 @@ use crate::{
         ConsensusManagerMetrics, ConsensusManagerTrait, ReplayWaiter, Running, RunningLockGuard,
     },
     consensus_validator::IotaTxValidator,
+    epoch_start_consensus_committee::EpochStartConsensusCommittee,
     starfish_adapter::LazyStarfishClient,
 };
